+++ examples/src/main/java/com/mapbox/navigation/examples/core/MapboxVoiceActivity.kt
@@ -110,7 +110,6 @@ class MapboxVoiceActivity : AppCompatActivity(), OnMapLongClickListener {
             .build()
         MapboxVoiceInstructionsPlayer(
             this,
-            getMapboxAccessTokenFromResources(),
             Locale.US.toLanguageTag(),
             options
         )