+++ libnavui-voice/src/test/java/com/mapbox/navigation/ui/voice/api/MapboxVoiceInstructionsPlayerTest.kt
@@ -74,14 +74,12 @@ class MapboxVoiceInstructionsPlayerTest {
     @Test
     fun updateLanguage() {
         val newLanguage = "IT"
-        val anyAccessToken = "pk.123"
         val anyLanguage = Locale.US.language
         val mockedFilePlayer: VoiceInstructionsFilePlayer = mockk(relaxed = true)
         val mockedTextPlayer: VoiceInstructionsTextPlayer = mockk(relaxed = true)
         every {
             retrieveVoiceInstructionsFilePlayer(
                 aMockedContext,
-                anyAccessToken,
                 mockedPlayerAttributes,
             )
         } returns mockedFilePlayer
@@ -95,7 +93,6 @@ class MapboxVoiceInstructionsPlayerTest {
         val mapboxVoiceInstructionsPlayer =
             MapboxVoiceInstructionsPlayer(
                 aMockedContext,
-                anyAccessToken,
                 anyLanguage,
                 mockedPlayerOptions
             )
@@ -107,7 +104,6 @@ class MapboxVoiceInstructionsPlayerTest {
 
     @Test
     fun `play VoiceInstructionsFilePlayer if file available`() {
-        val anyAccessToken = "pk.123"
         val anyLanguage = Locale.US.language
         val mockedFilePlayer: VoiceInstructionsFilePlayer = mockk()
         val mockedTextPlayer: VoiceInstructionsTextPlayer = mockk()
@@ -124,7 +120,6 @@ class MapboxVoiceInstructionsPlayerTest {
         every {
             retrieveVoiceInstructionsFilePlayer(
                 aMockedContext,
-                anyAccessToken,
                 mockedPlayerAttributes,
             )
         } returns mockedFilePlayer
@@ -139,7 +134,6 @@ class MapboxVoiceInstructionsPlayerTest {
         val mapboxVoiceInstructionsPlayer =
             MapboxVoiceInstructionsPlayer(
                 aMockedContext,
-                anyAccessToken,
                 anyLanguage,
                 mockedPlayerOptions
             )
@@ -179,7 +173,6 @@ class MapboxVoiceInstructionsPlayerTest {
 
     @Test
     fun `don't play VoiceInstruction if focus not granted`() {
-        val anyAccessToken = "pk.123"
         val anyLanguage = Locale.US.language
         val mockedFilePlayer: VoiceInstructionsFilePlayer = mockk()
         val mockedTextPlayer: VoiceInstructionsTextPlayer = mockk()
@@ -196,7 +189,6 @@ class MapboxVoiceInstructionsPlayerTest {
         every {
             retrieveVoiceInstructionsFilePlayer(
                 aMockedContext,
-                anyAccessToken,
                 mockedPlayerAttributes,
             )
         } returns mockedFilePlayer
@@ -211,7 +203,6 @@ class MapboxVoiceInstructionsPlayerTest {
         val mapboxVoiceInstructionsPlayer =
             MapboxVoiceInstructionsPlayer(
                 aMockedContext,
-                anyAccessToken,
                 anyLanguage,
                 mockedPlayerOptions
             )
@@ -251,7 +242,6 @@ class MapboxVoiceInstructionsPlayerTest {
 
     @Test
     fun `play VoiceInstructionsTextPlayer if file not available`() {
-        val anyAccessToken = "pk.123"
         val anyLanguage = Locale.US.language
         val mockedFilePlayer: VoiceInstructionsFilePlayer = mockk()
         val mockedTextPlayer: VoiceInstructionsTextPlayer = mockk()
@@ -268,7 +258,6 @@ class MapboxVoiceInstructionsPlayerTest {
         every {
             retrieveVoiceInstructionsFilePlayer(
                 aMockedContext,
-                anyAccessToken,
                 mockedPlayerAttributes,
             )
         } returns mockedFilePlayer
@@ -283,7 +272,6 @@ class MapboxVoiceInstructionsPlayerTest {
         val mapboxVoiceInstructionsPlayer =
             MapboxVoiceInstructionsPlayer(
                 aMockedContext,
-                anyAccessToken,
                 anyLanguage,
                 mockedPlayerOptions
             )
@@ -323,7 +311,6 @@ class MapboxVoiceInstructionsPlayerTest {
 
     @Test
     fun `announcements are played in order`() {
-        val anyAccessToken = "pk.123"
         val anyLanguage = Locale.US.language
         val mockedFilePlayer: VoiceInstructionsFilePlayer = mockk()
         val mockedTextPlayer: VoiceInstructionsTextPlayer = mockk()
@@ -346,7 +333,6 @@ class MapboxVoiceInstructionsPlayerTest {
         every {
             retrieveVoiceInstructionsFilePlayer(
                 aMockedContext,
-                anyAccessToken,
                 mockedPlayerAttributes,
             )
         } returns mockedFilePlayer
@@ -375,7 +361,6 @@ class MapboxVoiceInstructionsPlayerTest {
         val mapboxVoiceInstructionsPlayer =
             MapboxVoiceInstructionsPlayer(
                 aMockedContext,
-                anyAccessToken,
                 anyLanguage,
                 mockedPlayerOptions
             )
@@ -396,7 +381,6 @@ class MapboxVoiceInstructionsPlayerTest {
 
     @Test
     fun volume() {
-        val anyAccessToken = "pk.123"
         val anyLanguage = Locale.US.language
         val mockedFilePlayer: VoiceInstructionsFilePlayer = mockk()
         every { mockedFilePlayer.volume(any()) } just Runs
@@ -405,7 +389,6 @@ class MapboxVoiceInstructionsPlayerTest {
         every {
             retrieveVoiceInstructionsFilePlayer(
                 aMockedContext,
-                anyAccessToken,
                 mockedPlayerAttributes,
             )
         } returns mockedFilePlayer
@@ -419,7 +402,6 @@ class MapboxVoiceInstructionsPlayerTest {
         val mapboxVoiceInstructionsPlayer =
             MapboxVoiceInstructionsPlayer(
                 aMockedContext,
-                anyAccessToken,
                 anyLanguage,
                 mockedPlayerOptions
             )
@@ -438,7 +420,6 @@ class MapboxVoiceInstructionsPlayerTest {
 
     @Test(expected = IllegalArgumentException::class)
     fun `invalid volume less than min (0)`() {
-        val anyAccessToken = "pk.123"
         val anyLanguage = Locale.US.language
         val mockedFilePlayer: VoiceInstructionsFilePlayer = mockk()
         every { mockedFilePlayer.volume(any()) } just Runs
@@ -447,7 +428,6 @@ class MapboxVoiceInstructionsPlayerTest {
         every {
             retrieveVoiceInstructionsFilePlayer(
                 aMockedContext,
-                anyAccessToken,
                 mockedPlayerAttributes,
             )
         } returns mockedFilePlayer
@@ -461,7 +441,6 @@ class MapboxVoiceInstructionsPlayerTest {
         val mapboxVoiceInstructionsPlayer =
             MapboxVoiceInstructionsPlayer(
                 aMockedContext,
-                anyAccessToken,
                 anyLanguage,
                 mockedPlayerOptions
             )
@@ -473,7 +452,6 @@ class MapboxVoiceInstructionsPlayerTest {
 
     @Test(expected = IllegalArgumentException::class)
     fun `invalid volume greater than max (1)`() {
-        val anyAccessToken = "pk.123"
         val anyLanguage = Locale.US.language
         val mockedFilePlayer: VoiceInstructionsFilePlayer = mockk()
         every { mockedFilePlayer.volume(any()) } just Runs
@@ -482,7 +460,6 @@ class MapboxVoiceInstructionsPlayerTest {
         every {
             retrieveVoiceInstructionsFilePlayer(
                 aMockedContext,
-                anyAccessToken,
                 mockedPlayerAttributes,
             )
         } returns mockedFilePlayer
@@ -496,7 +473,6 @@ class MapboxVoiceInstructionsPlayerTest {
         val mapboxVoiceInstructionsPlayer =
             MapboxVoiceInstructionsPlayer(
                 aMockedContext,
-                anyAccessToken,
                 anyLanguage,
                 mockedPlayerOptions
             )
@@ -508,7 +484,6 @@ class MapboxVoiceInstructionsPlayerTest {
 
     @Test
     fun clear() {
-        val anyAccessToken = "pk.123"
         val anyLanguage = Locale.US.language
         val mockedFilePlayer: VoiceInstructionsFilePlayer = mockk()
         every { mockedFilePlayer.clear() } just Runs
@@ -517,7 +492,6 @@ class MapboxVoiceInstructionsPlayerTest {
         every {
             retrieveVoiceInstructionsFilePlayer(
                 aMockedContext,
-                anyAccessToken,
                 mockedPlayerAttributes,
             )
         } returns mockedFilePlayer
@@ -531,7 +505,6 @@ class MapboxVoiceInstructionsPlayerTest {
         val mapboxVoiceInstructionsPlayer =
             MapboxVoiceInstructionsPlayer(
                 aMockedContext,
-                anyAccessToken,
                 anyLanguage,
                 mockedPlayerOptions
             )
@@ -551,7 +524,6 @@ class MapboxVoiceInstructionsPlayerTest {
 
     @Test
     fun shutdown() {
-        val anyAccessToken = "pk.123"
         val anyLanguage = Locale.US.language
         val mockedFilePlayer: VoiceInstructionsFilePlayer = mockk()
         every { mockedFilePlayer.shutdown() } just Runs
@@ -560,7 +532,6 @@ class MapboxVoiceInstructionsPlayerTest {
         every {
             retrieveVoiceInstructionsFilePlayer(
                 aMockedContext,
-                anyAccessToken,
                 mockedPlayerAttributes,
             )
         } returns mockedFilePlayer
@@ -574,7 +545,6 @@ class MapboxVoiceInstructionsPlayerTest {
         val mapboxVoiceInstructionsPlayer =
             MapboxVoiceInstructionsPlayer(
                 aMockedContext,
-                anyAccessToken,
                 anyLanguage,
                 mockedPlayerOptions
             )
@@ -594,13 +564,11 @@ class MapboxVoiceInstructionsPlayerTest {
 
     @Test
     fun `request audio focus when play and abandon focus when done`() {
-        val anyAccessToken = "pk.123"
         val anyLanguage = Locale.US.language
         val mockedAnnouncement: SpeechAnnouncement = mockk(relaxed = true)
         val mapboxVoiceInstructionsPlayer =
             MapboxVoiceInstructionsPlayer(
                 aMockedContext,
-                anyAccessToken,
                 anyLanguage,
                 mockedPlayerOptions
             )
@@ -652,7 +620,6 @@ class MapboxVoiceInstructionsPlayerTest {
 
         val sut = MapboxVoiceInstructionsPlayer(
             context = aMockedContext,
-            accessToken = "access token",
             language = Locale.US.language,
             options = options,
             audioFocusDelegate = mockedAudioFocusDelegate,
@@ -680,7 +647,7 @@ class MapboxVoiceInstructionsPlayerTest {
 
         val filePlayerCallback = slot<VoiceInstructionsPlayerCallback>()
         val textPlayerCallback = slot<VoiceInstructionsPlayerCallback>()
-        every { retrieveVoiceInstructionsFilePlayer(any(), any(), any()) } returns mockk {
+        every { retrieveVoiceInstructionsFilePlayer(any(), any()) } returns mockk {
             every { play(any(), capture(filePlayerCallback)) } answers {
                 filePlayerCallbackAnswer(filePlayerCallback.captured)
             }