+++ libnavui-voice/src/test/java/com/mapbox/navigation/ui/voice/api/VoiceInstructionsTextPlayerTest.kt
@@ -424,7 +424,9 @@ class VoiceInstructionsTextPlayerTest {
             mockk<VoiceInstructionsPlayerCallback>(relaxUnitFun = true)
         val textPlayer =
             VoiceInstructionsTextPlayer(anyContext, language, anyPlayerAttributes)
-        every { mockedTextToSpeech.isLanguageAvailable(Locale(newLanguage)) } returns LANG_NOT_SUPPORTED
+        every {
+            mockedTextToSpeech.isLanguageAvailable(Locale(newLanguage))
+        } returns LANG_NOT_SUPPORTED
         invokeOnInitListener(TextToSpeech.SUCCESS)
         clearMocks(mockedTextToSpeech, answers = false)